move Main component to dir Main
