+++ project/src/components/main/main.tsx
@@ -1,4 +1,4 @@
-import SmallFilmCard from '../components/SmallFilmCard';
+import SmallFilmCard from '../SmallFilmCard';
 
 
 type MainScreenProps = {
